# [complex] torch.exp : match output against numpy