--- a/aten/src/ATen/cpu/vec256/vec256_complex_double.h
+++ b/aten/src/ATen/cpu/vec256/vec256_complex_double.h
@@ -218,15 +218,7 @@ public:
     AT_ERROR("not supported for complex numbers");
   }
   Vec256<c10::complex<double>> exp() const {
-    //exp(a + bi)
-    // = exp(a)*(cos(b) + sin(b)i)
-    auto exp = Sleef_expd4_u10(values);                               //exp(a)           exp(b)
-    exp = _mm256_blend_pd(exp, _mm256_permute_pd(exp, 0x05), 0x0A);   //exp(a)           exp(a)
-
-    auto sin_cos = Sleef_sincosd4_u10(values);                        //[sin(a), cos(a)] [sin(b), cos(b)]
-    auto cos_sin = _mm256_blend_pd(_mm256_permute_pd(sin_cos.y, 0x05),
-                                   sin_cos.x, 0x0A);                  //cos(b)           sin(b)
-    return _mm256_mul_pd(exp, cos_sin);
+    return map(std::exp);
   }
   Vec256<c10::complex<double>> expm1() const {
     AT_ERROR("not supported for complex numbers");