Fixes #48010