@kshitij12345 2020-11-16
I tried updating the SIMD code to cover the edges cases but couldn't cover all.
The special values are mentioned on https://en.cppreference.com/w/cpp/numeric/complex/exp

Unfortunately with using `map(std::exp)`, there is perf hit.
**Benchmark**

**Before**
```
Using pytorch 1.8.0a0+73e121d
10 / 10[-------- exp torch.complex64 --------]
                       |  torch.exp CPU
8 threads: ----------------------------
      (32,)            |        3.9    
      (32, 32)         |       15.1    
      (2, 16, 32)      |       12.5    
      (2, 16, 32, 32)  |       55.8    

Times are in microseconds (us).

[-------- exp torch.complex128 -------]
                       |  torch.exp CPU
8 threads: ----------------------------
      (32,)            |        3.7    
      (32, 32)         |       22.9    
      (2, 16, 32)      |       22.2    
      (2, 16, 32, 32)  |      100.6    

Times are in microseconds (us).
```

**After**
```
Using pytorch 1.8.0a0+46b252b
10 / 10[-------- exp torch.complex64 --------]
                       |  torch.exp CPU
8 threads: ----------------------------
      (32,)            |        4.4    
      (32, 32)         |       43.9    
      (2, 16, 32)      |       43.9    
      (2, 16, 32, 32)  |      182.8    

Times are in microseconds (us).

[-------- exp torch.complex128 -------]
                       |  torch.exp CPU
8 threads: ----------------------------
      (32,)            |        4.6    
      (32, 32)         |       60.1    
      (2, 16, 32)      |       48.6    
      (2, 16, 32, 32)  |      238.2    

Times are in microseconds (us).
```

Benchmark Code
```python
import torch
import itertools
import time
from torch.utils.benchmark import Timer
from torch.utils.benchmark import Compare
import sys

print('Using pytorch %s' % (torch.__version__))

shapes = [(32,), (32, 32), (2, 16, 32), (2, 16, 32, 32)]
results = []
num_threads = 8
repeats = 10

for dtype in [torch.cfloat, torch.cdouble]:
    for mat1_shape in shapes:
        mat1 = torch.rand(*mat1_shape, dtype=dtype, device='cpu')

        tasks = [("torch.exp(mat1)", "torch.exp CPU")]
        timers = [Timer(stmt=stmt, num_threads=num_threads, label=f"exp {dtype}", sub_label=f"{(mat1_shape)}", description=label, globals=globals()) for stmt, label in tasks]

        for i, timer in enumerate(timers * repeats):
            results.append(
                timer.blocked_autorange()
            )
            print(f"\r{i + 1} / {len(timers) * repeats}", end="")
            sys.stdout.flush()
comparison = Compare(results)
comparison.print()
```