@kshitij12345 2020-11-16
Reason for bfloat16 skip
```python
>>> import torch
>>> a = torch.tensor(6.7500)
>>> a
tensor(6.7500)
>>> torch.exp(a)
tensor(854.0588)
>>> torch.exp(a).to(torch.bfloat16)
tensor(856., dtype=torch.bfloat16)
```