@kshitij12345 2020-11-20
Bug in MSVC STL, 

Snippet from the issue,
```
exp(complex(inf, inf))
  Expected: (inf,nan)
    Actual: (inf,inf)
```
https://github.com/microsoft/STL/issues/887