@mruberry 2020-11-27
Hey @kshitij12345!

The latest attempt at fixing the vectorized code path is impressive. I'm curious if they'll actually be faster than the non-vectorized path and look forward to hearing your analysis.

As for errors in Windows and other compilers, we can skip tests or separate tests into "easy" and "hard" categories. For example, maybe on Linux exp is given extremal values but not on Windows or Mac. Then we can document the toolchain discrepancy with a warning in the documentation.