@kshitij12345 2020-11-27
> I'm curious if they'll actually be faster than the non-vectorized path and look forward to hearing your analysis.

The benchmark comment is updated with the timings for this new version and using map with std::exp.
https://github.com/pytorch/pytorch/pull/48014#issuecomment-728078206

The SIMD implementation is much faster than using map.