@kshitij12345 2022-04-27
I think the consensus was to have incorrect result for extremals if there is a perf regression. Closing this PR as it adds around `14%` perf regression for `complex64` and more for `complex128`.